requirements.txt
Django==1.7.3
django-allauth==0.19.0
django-appconf==0.6
django-extensions==1.4.9
django-polymorphic==0.6.1
oauthlib==0.7.2
python-openid==2.2.5
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
